create nav section
